+++ CHANGELOG.md
@@ -6,8 +6,13 @@ Version 3-SNAPSHOT
 
 CHANGED:
 
+- update to jdk version 21
 - rename of module to new name 'io.github.astrapisixtynine.gradle.migration.data'
 - major version to 3.*
+- update of gradle-plugin dependency with id 'io.freefair.lombok' to new patch version 8.10
+- update of gradle-plugin dependency with id 'com.diffplug.spotless' to new beta version 7.0.0.BETA2
+- update of dependency silly-collections to new major version 28
+- update of test dependency commons-lang3 to new version 3.17.0
 
 Version 2.4
 -------------